Add working tests ✨ reusing ellipse fit function created before - required refactoring
