+++ tests/test_integration/test_rotation_out_of_plane.py
@@ -1,14 +1,17 @@
+from pathlib import Path
 from typing import Optional, Tuple
 
 import numpy as np
 import pytest
 from matplotlib import pyplot as plt
 from skimage.draw import ellipse_perimeter
-from skimage.feature import canny
-from skimage.transform import hough_ellipse
 
+from derotation.analysis.fit_ellipse import (
+    fit_ellipse_to_points,
+    plot_ellipse_fit_and_centers,
+)
 from derotation.simulate.line_scanning_microscope import Rotator
-from tests.test_integration.test_finding_center_of_rotation_by_joining_two_pipelines import (
+from tests.test_integration.test_finding_center_of_rotation import (
     create_image_stack,
     create_rotation_angles,
     create_sample_image_with_two_cells,
@@ -42,7 +45,9 @@ def rotate_image_stack(
         Original image stack, rotated image stack, Rotator instance,
         and number of frames.
     """
-    cells = create_sample_image_with_two_cells(lines_per_frame=100)
+    cells = create_sample_image_with_two_cells(
+        lines_per_frame=100, second_cell=False, radius=1
+    )
     cells = np.pad(cells, ((pad, pad), (pad, pad)), mode="constant")
     cells[cells == 0] = 80
 
@@ -135,23 +140,23 @@ def make_plot(
         (25, None),
         (25, 45),
         (25, 90),
-        (25, 135),
         (40, None),
         (40, 45),
         (40, 90),
-        (40, 135),
     ],
 )
 def test_max_projection(
     plane_angle: int,
-    expected_orientation: Optional[int],
+    exp_orientation: Optional[int],
     atol: int = 15,
 ) -> None:
     """
     Tests if the max projection of an image stack fits an ellipse,
     and verify orientation and major/minor axes are close to expected values:
-    - if plane_angle is 0, the major and minor axes should be close to each other;
-    - if expected_orientation is not None, the orientation should be close to it.
+    - if plane_angle is 0, the major and minor axes should be close to each
+      other;
+    - if expected_orientation is not None, the orientation should be close to
+      it.
     - if expected_orientation is None, the orientation should be close to 0.
 
     Giving by default an allowed tolerance of 15 degrees as I care about
@@ -167,89 +172,95 @@ def test_max_projection(
         Allowed tolerance for orientation, by default 5.
     """
     _, rotated_image_stack, *_ = rotate_image_stack(
-        plane_angle=plane_angle, orientation=expected_orientation
+        plane_angle=plane_angle, orientation=exp_orientation
     )
 
-    max_projection = rotated_image_stack.max(axis=0)
-    edges = canny(max_projection, sigma=7, low_threshold=30, high_threshold=50)
-    result = hough_ellipse(
-        edges, accuracy=9, threshold=110, min_size=35, max_size=100
+    # get the location of the brightest pixel in any frame as a center
+    centers = np.array(
+        [
+            np.unravel_index(np.argmax(frame), frame.shape)
+            for frame in rotated_image_stack
+        ]
     )
-    result.sort(order="accumulator")
+    #  invert centers to (x, y) format
+    centers = np.array([(x, y) for y, x in centers])
 
-    best = result[-1]
-    yc, xc, a, b, orientation = [int(best[i]) for i in range(1, len(best))]
+    xc, yc, a, b, orientation = fit_ellipse_to_points(centers)
 
-    #  orientation: Major axis orientation in clockwise direction as radians.
-    #  expected orientation is calculated frm the minor axis, so convert
-    orientation = np.abs(np.rad2deg(orientation))
-
-    if expected_orientation is None and plane_angle == 0:
+    if exp_orientation is None and plane_angle == 0:
         #  lower tolerance for the major and minor axes
         if not np.allclose(a, b, atol=atol - 5):
-            plot_max_projection_with_rotation_out_of_plane(
-                max_projection,
-                edges,
-                yc,
-                xc,
-                a,
-                b,
-                orientation,
-                f"debug/error_{plane_angle}_{expected_orientation}.png",
+            plot_ellipse_fit_and_centers(
+                image_stack=rotated_image_stack,
+                centers=centers,
+                center_x=xc,
+                center_y=yc,
+                a=a,
+                b=b,
+                theta=orientation,
+                debug_plots_folder=Path("debug/"),
+                saving_name=f"ellipse_fit_{plane_angle}_{exp_orientation}.png",
             )
             assert (
                 False
-            ), f"Major and minor axes should be close to each other, instead got {a} and {b}"
-    elif expected_orientation is not None:
-        if not np.allclose(orientation, expected_orientation, atol=atol):
-            plot_max_projection_with_rotation_out_of_plane(
-                max_projection,
-                edges,
-                yc,
-                xc,
-                a,
-                b,
-                orientation,
-                f"debug/error_{plane_angle}_{expected_orientation}.png",
+            ), f"Major and minor axes should be close, instead got {a} and {b}"
+    elif exp_orientation is not None:
+        #  Major axis orientation in clockwise direction as radians.
+        exp_orientation = np.deg2rad(exp_orientation)
+        if not np.allclose(orientation, exp_orientation, atol=atol):
+            plot_ellipse_fit_and_centers(
+                image_stack=rotated_image_stack,
+                centers=centers,
+                center_x=xc,
+                center_y=yc,
+                a=a,
+                b=b,
+                theta=orientation,
+                debug_plots_folder=Path("debug/"),
+                saving_name=f"ellipse_fit_{plane_angle}_{exp_orientation}.png",
             )
-            assert (
-                False
-            ), f"Orientation should be close to {expected_orientation}, instead got {orientation}"
+            assert False, (
+                f"Orientation should be close to {exp_orientation}, "
+                f"instead got {orientation}"
+            )
     else:
-        # check that the major axis (a) differs from the minor axis (b) as calculated with cosine
+        # check that the major axis (a) differs from the minor axis (b)
+        # as calculated with cosine
         decrease = np.cos(np.deg2rad(plane_angle))
         expected_b = a - a * decrease
         #  lower tolerance for the major and minor axes
         if np.allclose(b, expected_b, atol=atol - 5):
-            plot_max_projection_with_rotation_out_of_plane(
-                max_projection,
-                edges,
-                yc,
-                xc,
-                a,
-                b,
-                orientation,
-                f"debug/error_{plane_angle}_{expected_orientation}.png",
+            plot_ellipse_fit_and_centers(
+                image_stack=rotated_image_stack,
+                centers=centers,
+                center_x=xc,
+                center_y=yc,
+                a=a,
+                b=b,
+                theta=orientation,
+                debug_plots_folder=Path("debug/"),
+                saving_name=f"ellipse_fit_{plane_angle}_{exp_orientation}.png",
             )
-            assert (
-                False
-            ), f"Difference between major and minor axes should be close to {expected_b}, instead got {b}"
+            assert False, (
+                f"Difference between major and minor axes should be close to "
+                f"{expected_b}, instead got {b}"
+            )
 
 
 def plot_max_projection_with_rotation_out_of_plane(
-    max_projection, edges, yc, xc, a, b, o, title
+    max_projection, yc, xc, a, b, o, title
 ):
     #  plot for debugging purposes
 
-    fig, ax = plt.subplots(1, 3, figsize=(10, 5))
+    fig, ax = plt.subplots(1, 2, figsize=(10, 5))
     ax[0].imshow(max_projection, cmap="gray")
     ax[0].set_title("Max projection")
     ax[0].axis("off")
 
-    #  plot edges
-    ax[1].imshow(edges, cmap="gray")
-    ax[1].set_title("Edges")
-    ax[1].axis("off")
+    # #  plot edges
+    # ax[1].imshow(edges, cmap="gray")
+    # ax[1].set_title("Edges")
+    # ax[1].axis("off")
 
     print(f"Ellipse center: ({yc}, {xc}), a: {a}, b: {b}, orientation: {o}")
 
@@ -259,9 +270,9 @@ def plot_max_projection_with_rotation_out_of_plane(
     empty_image[cy, cx] = 255
 
     #  show ellipse
-    ax[2].imshow(empty_image, cmap="gray")
-    ax[2].set_title("Ellipse")
-    ax[2].axis("off")
+    ax[1].imshow(empty_image, cmap="gray")
+    ax[1].set_title("Fitted ellipse")
+    ax[1].axis("off")
 
     plt.savefig(title)
